Implement world maps system for automatic map transitions
