+++ packages/tiledmap/src/index.ts
@@ -53,6 +53,3 @@ export function provideTiledMap(options: {
     }),
   ]);
 }
-
-// Export world maps functionality
-export { WorldMapsManager, RpgTiledWorldMap, WorldMapConfig } from './world-maps';
